requirements.txt
torch>=1.9.0
torchvision>=0.10.0
onnx>=1.10.0
onnxruntime>=1.9.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.3.0
mediapipe>=0.8.0
gdown>=4.0.0
PyYAML>=5.4.0
scipy>=1.7.0
tqdm>=4.62.0
ultralytics>=8.0.0 